requirements.txt
flask>=3.0.0
python-dotenv>=1.0.0
requests>=2.31.0
Pillow>=10.0.0
schedule>=1.2.0
google-generativeai>=0.7.0
beautifulsoup4>=4.12.0
